add comment to password and cpassword if statement

+++ server.php
@@ -8,7 +8,7 @@ if(isset($_POST['register-submit']))
     $phone = mysqli_real_escape_string($db_connect, $_POST['member-phone']);
     $password = mysqli_real_escape_string($db_connect, $_POST['member-password']);
     $cpassword = mysqli_real_escape_string($db_connect, $_POST['confirm-password']);
-    if(mysqli_num_rows($cpassword == $password))
+    if(mysqli_num_rows($cpassword == $password))//check if the password and confirm passsword match
     {
         echo '<script>alert("Profile created succesfully!!!")</script>';
         header ('location: login.php');
